+++ package.json
@@ -16,7 +16,8 @@
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
-    "swiper": "^11.1.14"
+    "swiper": "^11.1.14",
+    "workbox-build": "^7.1.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
