+++ src/pages/Profile/index.tsx
@@ -1,5 +1,11 @@
+import Profile from '@/components/profile/Profile';
+
 function ProfilePage() {
-  return <div className='pb-6 overflow-hidden'>마이페이지</div>;
+  return (
+    <div className='pb-6 overflow-hidden'>
+      <Profile />
+    </div>
+  );
 }
 
 export default ProfilePage;
